+++ crdtree/src/types/Types.ts
@@ -1,7 +1,6 @@
 import {CRDTreeTransport} from "../CRDTree";
 import {FrontendPrimitive} from "./Primitive";
 
-export type Timestamp = {clock: number, pid: string};
 export type ID = `${string}@${number}` | HEAD | ROOT | ROOT_PARENT;
 export type Index = number | string;
 export type ROOT = "ROOT";
